Split render data generator into its own module, removed calendar settings from renderer
