+++ resources/js/app.js
@@ -10,4 +10,5 @@ window.CalendarClock = require('./clock')
 window.RandomCalendar = require('./random-calendar')
 window.CalendarPresets = require('./calendar-presets')
 window.Perms = require('./perms');
+window.RenderDataGenerator = require('./render-data-generator')
 window.CalendarRenderer = require('./calendar-renderer')
