+++ resources/js/calendar/calendar_inputs_edit.js
@@ -4196,6 +4196,14 @@ function error_check(parent, rebuild){
 
 	}else{
 
+		RenderDataGenerator.create_render_data().then(
+			function(result){
+				window.dispatchEvent(new CustomEvent('render-data-change', {detail: result}));
+			}, function(err){
+				$.notify(err);
+			}
+		);
+
 		if(parent !== undefined && (parent === "seasons")){
 			rebuild_climate();
 		}else{
