+++ resources/js/webworkers/worker_calendar.js
@@ -2078,12 +2078,6 @@ var calendar_builder = {
 
 		climate_generator = new Climate(this.data.epochs, this.static_data, this.dynamic_data, calendar_start_epoch, calendar_end_epoch);
 		this.data.epochs = climate_generator.generate();
-		
-		let render_data = this.create_render_data({
-			start_epoch: calendar_start_epoch,
-			week_day: calendar_first_week_day,
-			year_day: calendar_year_day
-		});
 
 		if(debug || debugtext){
 
@@ -2136,345 +2130,10 @@ var calendar_builder = {
 			epoch_data: this.data.epochs,
 			processed_seasons: climate_generator.process_seasons,
 			processed_weather: climate_generator.process_weather,
-			render_data: render_data
+			timespans_to_build: this.calendar_list.timespans_to_build
 		}
-
-	},
-
-	get_weather_icon: function(weather){
-
-		if(weather){
-			if(weather.clouds == "Clear"){
-				if(weather.feature == "Fog"){
-					return `wi wi-fog`;
-				}else{
-					return `wi wi-day-sunny`;
-				}
-			}else{
-
-				if(weather.precipitation.key == "None"){
-
-					if(weather.clouds == "A few clouds"){
-						return `wi wi-day-cloudy`;
-					}else if(weather.clouds == "Mostly cloudy"){
-						return `wi wi-cloud`;
-					}else{
-						return `wi wi-cloudy`;
-					}
-
-				}else{
-
-					if(weather.temperature.metric.actual > 0){
-
-						if(weather.precipitation.actual > 0.375){
-							if(weather.feature == "Lightning"){
-								return `wi wi-thunderstorm`;
-							}else{
-								return `wi wi-rain`;
-							}
-						}else {
-							if(weather.feature == "Lightning"){
-								return `wi wi-storm-showers`;
-							}else{
-								if(weather.feature == "Fog" && weather.precipitation.actual < 0.375){
-									return `wi wi-fog`;
-								}
-								return `wi wi-showers`;
-							}
-						}
-
-					}else{
-
-						if(weather.feature == "Hail"){
-							return `wi wi-hail`;
-						}else{
-							return `wi wi-snow`;
-						}
-
-					}
-				}
-			}
-		}
-
-		return "";
-
-	},
-
-	get_day_data: function(epoch){
-	
-		let epoch_data = this.data.epochs[epoch];
-
-		let text = ""
-		if(epoch_data.leap_day !== undefined){
-			let index = epoch_data.leap_day;
-			text = this.static_data.year_data.leap_days[index].name;
-		}
-
-		let season_color = epoch_data.season ? (epoch_data.season.color !== undefined ? epoch_data.season.color : null) : null;
-		let weather_icon = epoch_data.weather ? this.get_weather_icon(epoch_data.weather) : null;
-
-		let moons = [];
-		for(let moon_index = 0; moon_index < this.static_data.moons.length; moon_index++){
-			let moon = this.static_data.moons[moon_index];
-			moons.push({
-				"index": moon_index,
-				"name": moon.name,
-				"phase": moon_phases[moon.granularity][data.moon_phase[moon_index]],
-				"image": svg_moon_shadows[Math.floor((svg_moon_shadows.length/moon.granularity)*epoch_data.moon_phase[moon_index])]
-			})
-		}
-
-		return {
-			"number": `${epoch_data.day}`,
-			"text": text,
-			"type": "day",
-			"epoch": epoch,
-			"year_day": epoch_data.year_day,
-			"weather_icon": weather_icon,
-			"season_color": season_color,
-			"events": [],
-			"moons": moons
-		};
-	},
-
-	overflow: function(){
-		return {
-			"number": ``,
-			"text": "",
-			"type": "overflow",
-			"epoch": false,
-			"year_day": false,
-			"weather_icon": false,
-			"season_color": false,
-			"events": [],
-			"moons": []
-		};
-	},
-
-	create_render_data: function(data){
-
-		let render_data = {
-			"current_epoch": this.dynamic_data.epoch,
-			"preview_epoch": this.dynamic_data.epoch,
-			"render_style": this.static_data.settings.layout,
-			"timespans": [],
-			"event_epochs": {}
-		}
-
-		let indexes = Object.keys(this.calendar_list.timespans_to_build)
-		let length = indexes.length
-
-		let epoch = data.start_epoch;
-		let week_day = data.week_day;
-
-		for(var index = 0; index < length; index++){
-
-			let timespan = this.calendar_list.timespans_to_build[indexes[index]];
-
-			var filtered_leap_days_beforestart = timespan.leap_days.filter(function(features){
-				return features.intercalary && features.day === 0;
-			});
-
-			if(filtered_leap_days_beforestart.length > 0){
-					
-				let timespan_data = {
-					"title": "",
-					"show_title": false,
-					"number": index+1,
-					"weekdays": this.static_data.year_data.global_week,
-					"show_weekdays": false,
-					"days": [[]]
-				}
-
-				let weekday_number = 1;
-				let week_index = 0;
-
-				for(var leap_day_index in filtered_leap_days_beforestart){
-
-					let day_data = this.get_day_data(epoch);
-					timespan_data.days[timespan_data.days.length-1].push(day_data);
-					render_data.event_epochs[epoch] = day_data.events;
-
-					weekday_number++;
-					epoch++;
-
-					if(weekday_number > timespan.week.length){
-						weekday_number = 1;
-						week_index++;
-						timespan_data.days.push([]);
-					}
-				}
-
-				for(weekday_number; weekday_number <= this.static_data.year_data.global_week.length; weekday_number++){
-					timespan_data.days[week_index].push(this.overflow());
-				}
-
-				render_data.timespans.push(timespan_data);
-
-			}
-
-			let show_months = timespan.type === "month";
-
-			let timespan_data = {
-				"title": timespan.name,
-				"show_title": true,
-				"number": index+1,
-				"weekdays": timespan.week,
-				"show_weekdays": show_months,
-				"days": [[]]
-			}
-
-			if(!this.static_data.year_data.overflow){
-				week_day = 1;
-			}
-
-			for(let day_number = 1; day_number <= timespan.length;){
-				
-				if(timespan_data.days[timespan_data.days.length-1].length != 0){
-					timespan_data.days.push([])
-				}
-
-				for(let weekday_number = 1; weekday_number <= timespan.week.length; weekday_number++){
-
-					if(week_day > weekday_number && show_months){
-
-						timespan_data.days[timespan_data.days.length-1].push(this.overflow());
-						
-					}else if(day_number <= timespan.length){
-
-						let day_data = this.get_day_data(epoch);
-						timespan_data.days[timespan_data.days.length-1].push(day_data);
-						render_data.event_epochs[epoch] = day_data.events;
-
-						epoch++;
-
-						filtered_leap_days = timespan.leap_days.filter(function(leap_day){
-							return leap_day.intercalary && leap_day.day === day_number && leap_day.day !== timespan.length;
-						});
-
-						if(filtered_leap_days.length > 0){
-
-							for(let internal_weekday_number = week_day; internal_weekday_number < timespan.week.length; internal_weekday_number++){
-								timespan_data.days[timespan_data.days.length-1].push(this.overflow());
-							}
-
-							render_data.timespans.push(timespan_data);
-
-							timespan_data = {
-								"title": "",
-								"show_title": false,
-								"number": index+1,
-								"weekdays": this.static_data.year_data.global_week,
-								"show_weekdays": false,
-								"days": [[]]
-							}
-
-							let internal_weekday_number = 1;
-
-							for(var leap_day_index in filtered_leap_days){
-
-								let day_data = this.get_day_data(epoch);
-								timespan_data.days[timespan_data.days.length-1].push(day_data);
-								render_data.event_epochs[epoch] = day_data.events;
-
-								internal_weekday_number++;
-								epoch++;
-
-								if(internal_weekday_number > timespan.week.length){
-									internal_weekday_number = 1;
-									timespan_data.days.push([]);
-								}
-							}
-
-							for(internal_weekday_number; internal_weekday_number <= this.static_data.year_data.global_week.length; internal_weekday_number++){
-								timespan_data.days[timespan_data.days.length-1].push(this.overflow());
-							}
-
-							render_data.timespans.push(timespan_data);
-
-							timespan_data = {
-								"title": timespan.name,
-								"show_title": true,
-								"number": index+1,
-								"weekdays": timespan.week,
-								"show_weekdays": timespan.type === "month",
-								"days": [[]]
-							}
-
-							if(week_day != timespan.week.length){
-								for(let internal_weekday_number = 0; internal_weekday_number < week_day; internal_weekday_number++){
-									timespan_data.days[timespan_data.days.length-1].push(this.overflow());
-								}
-							}
-
-						}
-
-						day_number++;
-
-						if(show_months){
-							week_day++;
-							
-							if(week_day > timespan.week.length){
-								week_day = 1;
-							}
-						}
-
-					}else{
-
-						timespan_data.days[timespan_data.days.length-1].push(this.overflow());
-
-					}
-				}
-			}
-
-			render_data.timespans.push(timespan_data);
-
-			var filtered_leap_days_end = timespan.leap_days.filter(function(features){
-				return features.intercalary && features.day === timespan.length;
-			});
-
-			if(filtered_leap_days_end.length > 0){
-					
-				let timespan_data = {
-					"title": "",
-					"show_title": false,
-					"number": index+1,
-					"weekdays": this.static_data.year_data.global_week,
-					"show_weekdays": false,
-					"days": [[]]
-				}
-
-				let weekday_number = 1;
-				let week_index = 0;
-
-				for(var leap_day_index in filtered_leap_days_end){
-
-					timespan_data.days[week_index].push(this.get_day_data(epoch));
-
-					weekday_number++;
-					epoch++;
-
-					if(weekday_number > timespan.week.length){
-						weekday_number = 1;
-						week_index++;
-						timespan_data.days.push([]);
-					}
-				}
-
-				for(weekday_number; weekday_number <= this.static_data.year_data.global_week.length; weekday_number++){
-					timespan_data.days[week_index].push(this.overflow());
-				}
-
-				render_data.timespans.push(timespan_data);
-
-			}
-
-		}
-		
-		return render_data;
 
 	}
-
 
 }
