+++ resources/views/layouts/calendar.blade.php
@@ -1,7 +1,7 @@
 <div id='calendar' x-data="CalendarRenderer">
 
     <template
-        @calendar-change.window="
+        @render-data-change.window="
             pre_render();
             load_calendar($event);
             $nextTick(() => { post_render() });
@@ -12,7 +12,6 @@
             $nextTick(() => { post_event_load() });
         "
         @update-epochs.window="update_epochs"
-        @render-settings-change.window="update_render_settings"
         x-if='loaded'
         x-for="timespan in render_data.timespans"
     >
@@ -21,12 +20,12 @@
 
             <div class='timespan_name'>
                 <span x-text='timespan.title'></span>
-                <span class='timespan_number' x-show="render_settings.add_month_number" x-text='[" - " + timespan.number]'></span>
+                <span class='timespan_number' x-show="timespan.number" x-text='[" - " + timespan.number]'></span>
             </div>
 
             <div class='timespan_row_container'>
 
-                <div x-show='timespan.show_weekdays && !render_settings.hide_weekdays' class='timespan_row_names'>
+                <div x-show='timespan.show_weekdays' class='timespan_row_names'>
                     <template x-for="weekday in timespan.weekdays">
                         <div class='week_day_name' x-text='weekday'></div>
                     </template>
@@ -38,28 +37,16 @@
                             <div :class="{
                                 'timespan_day': day.type == 'day',
                                 'timespan_overflow': day.type == 'overflow',
+                                'timespan_day empty_timespan_day': day.type == 'empty',
                                 'current_day': day.epoch == render_data.current_epoch,
                                 'preview_day': day.epoch == render_data.preview_epoch && render_data.preview_epoch != render_data.current_epoch,
-                                'empty_timespan_day': !render_settings.owner && day.type != 'overflow' && render_settings.only_reveal_today && day.epoch > render_data.current_epoch
                             }" :epoch="day.epoch">
                                 <div class='day_row'>
-                                    <div class='toprow left'>
+                                    <div class='toprow left' x-show="day.number">
                                         <div class='number' x-text='day.number'></div>
                                     </div>
                                     <div class='toprow center'>
-                                        <template x-if="
-                                            day.weather_icon
-                                            &&
-                                            (
-                                                render_settings.owner
-                                                ||
-                                                !(
-                                                    render_settings.hide_all_weather
-                                                    ||
-                                                    (render_settings.hide_future_weather && day.epoch > render_data.current_epoch)
-                                                )
-                                            )
-                                        ">
+                                        <template x-if="day.weather_icon">
                                             <div class='weather_popup'><i x-bind:class='day.weather_icon'></i></div>
                                         </template>
                                     </div>
@@ -67,18 +54,20 @@
                                         <div class='season_color'></div>
                                     </div>
                                 </div>
-                                <div class='day_row flex justify-content-center'>
-                                    <template x-for="moon in day.moons" x-if='render_settings.owner || !render_settings.hide_moons'>
-                                        <div class='moon_container protip'
-                                            :moon_id="moon.index"
-                                            data-pt-position="top"
-                                            data-pt-title=''>
-                                            <i class='wi wi-moon-full'></i>
-                                        </div>
+                                <div x-show="day.moons" class='day_row flex justify-content-center'>
+                                    <template x-if='day.moons'>
+                                        <template x-for="moon in day.moons">
+                                            <div class='moon_container protip'
+                                                :moon_id="moon.index"
+                                                data-pt-position="top"
+                                                data-pt-title=''>
+                                                <i class='wi wi-moon-full'></i>
+                                            </div>
+                                        </template>
                                     </template>
                                 </div>
-                                <div class="day_row" x-show="day.events.length">
-                                    <template x-if="render_settings.owner || !render_settings.hide_events">
+                                <div class="day_row">
+                                    <template x-if="day.events">
                                         <div class="event_container">
                                             <template x-for="calendar_event in day.events">
                                                 <div class="event" :class="calendar_event.class" x-text="calendar_event.name" :event_id="calendar_event.id"></div>
@@ -89,8 +78,8 @@
                                 <div class="day_row flex-grow" x-show="day.type === 'day'">
                                     <button class="btn_create_event btn btn-success full" @click="create_event(day.epoch)" :epoch="day.epoch">Create event</button>
                                 </div>
-                                <div class="day_row" x-show="render_settings.add_year_day_number && day.year_day">
-                                    <div class="year_day" x-text="day.year_day"></div>
+                                <div class="day_row">
+                                    <div class="year_day" x-show="day.year_day" x-text="day.year_day"></div>
                                 </div>
                             </div>
                         </template>
